chore: Overhaul nix flake
This brings in a series of changes to make the nix flake more modular
using flake-parts. The crate has been changed to exist in a workspace
which can have many crates. Some cargo dependencies have been updated to
the latest versions. The flake now includes a set of checks on the
codebase including styling with treefmt, clippy, and more.